Revert "Moving unit tests from create pool to a different file from integrations;"
This reverts commit 96d353502d5b161ef2aa1ce612aea96a0fbf43cb.
