+++ test/v2/createPool/weighted.integration.test.ts
@@ -77,4 +77,81 @@ describe('Create Weighted Pool tests', () => {
         });
         expect(poolAddress).to.not.be.undefined;
     });
+
+    test('Wrong weights, expects error', async () => {
+        const tokens: CreatePoolV2WeightedInput['tokens'] = [
+            {
+                address: '0xba100000625a3754423978a60c9317c58a424e3d',
+                weight: parseEther(`${1 / 3}`),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+                weight: parseEther(`${1 / 3}`),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0x6b175474e89094c44da98b954eedeac495271d0f',
+                weight: parseEther(`${1 / 3}`),
+                rateProvider: zeroAddress,
+            },
+        ];
+        await expect(() =>
+            doCreatePool({
+                ...txInput,
+                createPoolInput: { ...createWeightedPoolInput, tokens },
+            }),
+        ).rejects.toThrowError('Weights must sum to 1e18');
+    });
+
+    test('Weight value 0, expects error', async () => {
+        const tokens: CreatePoolV2WeightedInput['tokens'] = [
+            {
+                address: '0xba100000625a3754423978a60c9317c58a424e3d',
+                weight: parseEther('0'),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+                weight: parseEther(`${1 / 2}`),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0x6b175474e89094c44da98b954eedeac495271d0f',
+                weight: parseEther(`${1 / 2}`),
+                rateProvider: zeroAddress,
+            },
+        ];
+        await expect(() =>
+            doCreatePool({
+                ...txInput,
+                createPoolInput: { ...createWeightedPoolInput, tokens },
+            }),
+        ).rejects.toThrowError('Weight cannot be 0');
+    });
+    test('Duplicate token addresses, expects error', async () => {
+        const tokens: CreatePoolV2WeightedInput['tokens'] = [
+            {
+                address: '0xba100000625a3754423978a60c9317c58a424e3d',
+                weight: parseEther(`${1 / 3}`),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0xba100000625a3754423978a60c9317c58a424e3d',
+                weight: parseEther(`${1 / 3}`),
+                rateProvider: zeroAddress,
+            },
+            {
+                address: '0x6b175474e89094c44da98b954eedeac495271d0f',
+                weight: parseEther(`${1 / 3 + 1e-16}`),
+                rateProvider: zeroAddress,
+            },
+        ];
+        await expect(() =>
+            doCreatePool({
+                ...txInput,
+                createPoolInput: { ...createWeightedPoolInput, tokens },
+            }),
+        ).rejects.toThrowError('Duplicate token addresses');
+    });
 });
